requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
aiohttp==3.8.1
persuader-technology-automata-core>=0.6.0
persuader-technology-automata-utilities>=0.2.3
persuader-technology-automata-logger>=0.0.5
persuader-technology-automata-auth>=0.1.6
persuader-technology-automata-redis>=0.3.8
persuader-technology-automata-timeseries>=0.0.6
persuader-technology-automata-missing-repository>=0.2.1
persuader-technology-automata-exchange-repository>=0.3.1
persuader-technology-automata-exchange-transform-repository>=0.2.0
persuader-technology-automata-trade-transform-repository>=0.1.2
persuader-technology-automata-process-repository>=0.2.1
